+++ circuit/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod builder;
 pub mod circuit;
 pub mod expr;
-pub mod prim;
+pub mod op;
 pub mod tables;
 pub mod types;
 
@@ -9,6 +9,6 @@ pub mod types;
 pub use builder::CircuitBuilder;
 pub use circuit::Circuit;
 pub use expr::{Expr, ExpressionGraph};
-pub use prim::{FakeMerklePrivateData, NonPrimitiveOp, NonPrimitiveOpPrivateData, Prim};
+pub use op::{FakeMerklePrivateData, NonPrimitiveOp, NonPrimitiveOpPrivateData, Prim};
 pub use tables::{CircuitRunner, FakeMerkleTrace, Traces};
 pub use types::{ExprId, NonPrimitiveOpId, WitnessAllocator, WitnessId};
